requirements.txt
asgiref==3.5.2
auth==0.5.3
bcrypt==4.0.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-oauth-toolkit==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
drf-yasg==1.21.4
eventlet==0.33.1
fabric==2.7.1
facebook-sdk==3.1.0
fake==0.8
falcon==3.1.0
google==3.0.0
google-api-core==2.10.1
google-auth==2.12.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
inflection==0.5.1
invoke==1.7.3
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
mongoengine==0.24.2
oauth2==1.9.0.post1
oauthlib==3.2.1
packaging==21.3
paramiko==2.11.0
pathlib2==2.3.7.post1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pymongo==4.2.0
PyNaCl==1.5.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
twitter==1.19.6
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
